Upgraded Gravity Forms to 1.8.6

+++ wp-content/plugins/gravityforms/entry_list.php
@@ -74,8 +74,13 @@ class GFEntryList{
             case "delete" :
                 check_admin_referer('gforms_entry_list', 'gforms_entry_list');
                 $lead_id = $_POST["action_argument"];
-                RGFormsModel::delete_lead($lead_id);
-                $update_message = __("Entry deleted.", "gravityforms");
+                if(GFCommon::current_user_can_any("gravityforms_delete_entries")){
+                    RGFormsModel::delete_lead($lead_id);
+                    $update_message = __("Entry deleted.", "gravityforms");
+                } else {
+                    $update_message = __("You don't have adequate permissions to delete entries.", "gravityforms");
+                }
+
             break;
 
             case "bulk":
@@ -89,8 +94,12 @@ class GFEntryList{
 
                 switch($bulk_action) {
                     case "delete":
-                        RGFormsModel::delete_leads($leads);
-                        $update_message = sprintf(__("%s deleted.", "gravityforms"), $entry_count);
+                        if(GFCommon::current_user_can_any("gravityforms_delete_entries")){
+                            RGFormsModel::delete_leads($leads);
+                            $update_message = sprintf(__("%s deleted.", "gravityforms"), $entry_count);
+                        } else {
+                            $update_message = __("You don't have adequate permissions to delete entries.", "gravityforms");
+                        }
                     break;
 
                     case "trash":
@@ -143,7 +152,9 @@ class GFEntryList{
         }
 
         if(rgpost("button_delete_permanently")){
-            RGFormsModel::delete_leads_by_form($form_id, $filter);
+            if(GFCommon::current_user_can_any("gravityforms_delete_entries")){
+                RGFormsModel::delete_leads_by_form($form_id, $filter);
+            }
         }
 
         $sort_field = empty($_GET["sort"]) ? 0 : $_GET["sort"];
@@ -905,7 +916,7 @@ class GFEntryList{
                         }
                         ?>
                         <th scope="col" align="right" width="50">
-                            <a title="<?php _e("click to select columns to display" , "gravityforms") ?>" href="<?php echo trailingslashit(site_url()) ?>?gf_page=select_columns&id=<?php echo $form_id ?>&TB_iframe=true&height=365&width=600" class="thickbox entries_edit_icon"><i class="fa fa-cog"></i></a>
+                            <a title="<?php _e("click to select columns to display" , "gravityforms") ?>" href="<?php echo trailingslashit(site_url(null, "admin")) ?>?gf_page=select_columns&id=<?php echo $form_id ?>&TB_iframe=true&height=365&width=600" class="thickbox entries_edit_icon"><i class="fa fa-cog"></i></a>
                         </th>
                     </tr>
                 </thead>
@@ -1076,7 +1087,9 @@ class GFEntryList{
                                         case "created_by" :
                                             if(!empty($value)){
                                                 $userdata = get_userdata($value);
-                                                $value = $userdata->user_login;
+                                                if(!empty($userdata)){
+                                                    $value = $userdata->user_login;
+                                                }
                                             }
                                         break;
